pyproject.toml
[project]
name = "ai-researcher"
version = "0.1.0"
description = "AI-powered research automation system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-xdist>=3.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "bandit>=1.7.0",
    "twine>=4.0.0",
    "codecov>=2.1.0"
]

[project.scripts]
ai-researcher = "src.main:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# Rule selection
select = [
    "F",      # Pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort (import sorting)
    "UP",     # pyupgrade (modern Python idioms)
    "B",      # flake8-bugbear
    "BLE",    # flake8-blind-except
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PGH",    # pygrep-hooks
    "PL",     # Pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "NPY",    # NumPy-specific rules
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
    "PLR0915", # Too many statements
    "TRY003",  # Avoid specifying long messages outside the exception class
    "EM101",   # Exception string literals (allow for simple cases)
    "EM102",   # Exception f-strings (allow for simple cases)
    "BLE001",  # Allow blind exception handling in config
    "ISC001",  # Implicitly concatenated string literals on one line (conflicts with formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "ARG001",  # Unused function arguments in tests
    "PLR2004", # Magic value comparison in tests
    "S101",    # Use of assert in tests
]

# MyPy per-file ignores
[[tool.mypy.overrides]]
module = [
    "tests.unit.test_pyproject_config",
    "tests.unit.test_precommit_config"
]
ignore_errors = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

# Per-module options
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "openai.*",
    "anthropic.*",
    "pandas.*",
    "rich.*",
    "yaml.*"
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", ".git"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection_process

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "network: Tests that require network access",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 80

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "pre-commit>=4.4.0",
    "pytest-cov>=7.0.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
    "types-setuptools>=80.9.0.20250822",
    "types-toml>=0.10.8.20240310",
]
